코드 수정, 회원탈퇴 기능 추가
