+++ src/main/java/com/wappenable/be/users/controller/AdminUserController.java
@@ -16,6 +16,7 @@ import com.wappenable.be.users.dto.request.RoleUpdateRequestDto;
 import com.wappenable.be.users.dto.response.RoleChangedMessage;
 import com.wappenable.be.users.dto.response.UserListDto;
 import com.wappenable.be.users.entity.Role;
+import com.wappenable.be.users.service.AdminUserService;
 import com.wappenable.be.users.service.UserService;
 
 import lombok.RequiredArgsConstructor;
@@ -26,14 +27,14 @@ import org.springframework.web.bind.annotation.GetMapping;
 @RequestMapping("/api/admin")
 @RequiredArgsConstructor
 public class AdminUserController {
-    private final UserService userService;
+    private final AdminUserService adminUserService;
     private final SimpMessagingTemplate messagingTemplate;
 
     // 50명 초기 목록 렌더링을 위해 필요
     @GetMapping("/users")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<UserListDto>> getUserList() {
-        List<UserListDto> users = userService.getAllUsers();
+        List<UserListDto> users = adminUserService.getAllUsers();
         return ResponseEntity.ok(users);
     }
 
@@ -42,7 +43,7 @@ public class AdminUserController {
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> updateUserRole(@PathVariable Long id,
                                                @RequestBody RoleUpdateRequestDto req) {
-        userService.updateUserRole(id, req.getRole());
+                                                adminUserService.updateUserRole(id, req.getRole());
 
         // WebSocket 실시간 알림 전송
         messagingTemplate.convertAndSend("/topic/users/role",
@@ -60,6 +61,5 @@ public class AdminUserController {
             .toList();
         return ResponseEntity.ok(roles);
     }
-    
 
 }
