+++ src/main/java/com/wappenable/be/users/service/UserService.java
@@ -2,6 +2,7 @@ package com.wappenable.be.users.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,13 +12,10 @@ import com.wappenable.be.global.security.jwt.JwtUtil;
 import com.wappenable.be.global.security.jwt.TokenResponse;
 import com.wappenable.be.users.dto.request.LoginRequest;
 import com.wappenable.be.users.dto.request.SignupRequest;
-import com.wappenable.be.users.dto.response.UserListDto;
 import com.wappenable.be.users.entity.Role;
 import com.wappenable.be.users.entity.User;
 import com.wappenable.be.users.repository.UserRepository;
 
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -64,30 +62,12 @@ public class UserService {
     }
 
     @Transactional
-    public void updateUserRole(Long userId, String roleName) {
-        User user = userRepository.findById(userId)
-            .orElseThrow(() -> new CustomException("사용자 없음",HttpStatus.BAD_REQUEST));
+    public void deleteCurrentUser() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
 
-        try {
-            Role newRole = Role.valueOf(roleName.toUpperCase());
-            user.setRole(newRole);
-        } catch (IllegalArgumentException e) {
-            throw new CustomException("잘못된 역할 값입니다", HttpStatus.BAD_REQUEST);
-        }
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new CustomException("사용자를 찾을 수 없습니다.", HttpStatus.UNAUTHORIZED));
+
+        userRepository.delete(user);
     }
-
-    // 전체 사용자 조회
-    @Transactional(readOnly = true)
-    public List<UserListDto> getAllUsers() {
-        return userRepository.findAll().stream()
-                .map(user -> new UserListDto(
-                        user.getId(),
-                        user.getEmail(),
-                        user.getNickname(),
-                        user.getRole().name()
-                ))
-                .collect(Collectors.toList());
-        }
-
-
 }
