+++ src/test/java/com/wappenable/be/service/UserServiceTest.java
@@ -71,89 +71,4 @@ class UserServiceTest {
 
         verify(userRepository, never()).save(any(User.class));
     }
-
-    @Test
-    void 역할_변경_성공() {
-        // given
-        Long userId = 1L;
-        User user = User.builder()
-                .id(userId)
-                .email("test@wappen.com")
-                .nickname("테스트")
-                .role(Role.USER)
-                .build();
-
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-
-        // when
-        userService.updateUserRole(userId, "ADMIN");
-
-        // then
-        verify(userRepository).findById(userId);
-        assertThat(user.getRole()).isEqualTo(Role.ADMIN);
-    }
-
-    @Test
-    void 역할_변경_실패_잘못된_역할() {
-        // given
-        Long userId = 1L;
-        User user = User.builder()
-                .id(userId)
-                .email("test@wappen.com")
-                .nickname("테스트")
-                .role(Role.USER)
-                .build();
-
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-
-        // when & then
-        assertThatThrownBy(() -> userService.updateUserRole(userId, "INVALID_ROLE"))
-                .isInstanceOf(CustomException.class)
-                .hasMessage("잘못된 역할 값입니다");
-    }
-
-    @Test
-    void 역할_변경_실패_없는_사용자() {
-        // given
-        Long userId = 99L;
-        when(userRepository.findById(userId)).thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> userService.updateUserRole(userId, "ADMIN"))
-                .isInstanceOf(CustomException.class)
-                .hasMessage("사용자 없음");
-    }
-
-    @Test
-    void 전체_사용자_조회_성공() {
-        // given
-        List<User> users = List.of(
-            User.builder()
-                .id(1L)
-                .email("user1@example.com")
-                .nickname("유저1")
-                .passwordHash("encoded1")
-                .role(Role.USER)
-                .build(),
-            User.builder()
-                .id(2L)
-                .email("admin@example.com")
-                .nickname("관리자")
-                .passwordHash("encoded2")
-                .role(Role.ADMIN)
-                .build()
-        );
-
-        when(userRepository.findAll()).thenReturn(users);
-
-        // when
-        var result = userService.getAllUsers();
-
-        // then
-        assertThat(result).hasSize(2);
-        assertThat(result.get(0).getEmail()).isEqualTo("user1@example.com");
-        assertThat(result.get(0).getRole()).isEqualTo("USER");
-        assertThat(result.get(1).getNickname()).isEqualTo("관리자");
-        assertThat(result.get(1).getRole()).isEqualTo("ADMIN");
-    }
 }
